pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supplyhub"
version = "0.1.0"
description = "Internal service for reconciling foreign trade documents"
readme = "SPEC.md"
requires-python = ">=3.11"
authors = [{ name = "SupplyHub Team" }]
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.23",
    "pydantic>=2.5",
    "pydantic-settings>=2.2",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.12",
    "celery>=5.3",
    "redis>=5.0",
    "httpx>=0.27",
    "python-multipart>=0.0.7",
    "jinja2>=3.1",
    "itsdangerous>=2.1",
    "PyMuPDF>=1.23",
    "python-docx>=1.0",
    "openpyxl>=3.1",
    "pillow>=10.0",
    "openai>=1.40"
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-asyncio>=0.21", "coverage>=7.3"]

[tool.setuptools.packages.find]
where = ["app"]

